Use match-case instead of long if-else statements
In this case a switch case (or match case in Python) is preferred. It
grabs the content of the status variable and jumps directly to the case
instead of trying out all possibilites in a row. This improves
performance of the statement.

Python version needs to be updated to 3.10 for the match-case to work.